pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "radiogenpdac"
version = "0.1.0"
description = "Cluster-ready PDAC radiogenomics framework scaffold."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "rich>=13.7",
  "scikit-learn>=1.5",
  "typer>=0.12",
]

[project.optional-dependencies]
train = [
  "einops>=0.8",
  "lightning>=2.5",
  "mlflow>=2.16",
  "monai>=1.5",
  "nibabel>=5.2",
  "SimpleITK>=2.3",
  "torch>=2.3",
  "torchmetrics>=1.4",
]
dev = [
  "pytest>=8.2",
  "ruff>=0.6",
]

[project.scripts]
radiogenpdac = "radiogenpdac.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
